scheduler: lightly refactor InterruptController
- Change the file names to more accurately describe their contents

- Split out the static configuration bits from the futex-adaptor bits
